Add royalty context to registerDerivativeIp
